Fix: header styled component refactoring
