+++ src/components/Header/utils.js
@@ -1,0 +1,8 @@
+const pathButtons = [
+  { path: '/reservation', text: '예약/구매내역' },
+  { path: '/likes', text: '위시리스트' },
+  { path: '/mydaily', text: '마이데일리' },
+  { path: '/current', text: '최근 본 상품' },
+];
+
+export { pathButtons };
